refactor: 完成项目大规模代码清理与规范重构
本次提交进行了全面的代码优化与重构工作：
1.  **枚举迁移**：将多处分散的枚举定义迁移至模型层，统一管理并移除冗余的enums.py文件
2.  **路由优化**：标准化API路由路径，移除多余的末尾斜杠，统一路由注册方式
3.  **类型安全改进**：将字符串类型的字段替换为强类型枚举，修复多处类型不匹配问题
4.  **软删除增强**：为社区别名等模型添加逻辑删除支持，优化查询过滤条件
5.  **依赖注入**：重构小区服务，引入依赖注入工厂函数
6.  **参数验证**：新增排序字段验证工具，修复SQL注入风险
7.  **加密增强**：新增敏感字段加密支持，实现自动加解密的数据库字段类型
8.  **工具类重构**：重构CSV导出、路径校验等工具函数，修复路径遍历安全漏洞
9.  **测试更新**：同步更新所有测试用例以适配接口变更
10. **配置完善**：新增前端URL、加密密钥等配置项，完善系统环境变量支持

+++ backend/schemas/project/renovation.py
@@ -73,28 +73,28 @@ class RenovationContractUpdate(BaseModel):
     actual_end_date: datetime | None = Field(None, description="实际竣工时间")
 
     # 硬装费用
-    hard_contract_amount: float | None = Field(None, description="硬装合同总金额")
+    hard_contract_amount: Decimal | None = Field(None, description="硬装合同总金额")
 
     # 支付节点
     payment_node_1: str | None = Field(None, max_length=100, description="第一笔款项支付节点")
-    payment_ratio_1: float | None = Field(None, ge=0, le=100, description="第一笔款项支付比例")
+    payment_ratio_1: Decimal | None = Field(None, ge=0, le=100, description="第一笔款项支付比例")
     payment_node_2: str | None = Field(None, max_length=100, description="第二笔款项支付节点")
-    payment_ratio_2: float | None = Field(None, ge=0, le=100, description="第二笔款项支付比例")
+    payment_ratio_2: Decimal | None = Field(None, ge=0, le=100, description="第二笔款项支付比例")
     payment_node_3: str | None = Field(None, max_length=100, description="第三笔款项支付节点")
-    payment_ratio_3: float | None = Field(None, ge=0, le=100, description="第三笔款项支付比例")
+    payment_ratio_3: Decimal | None = Field(None, ge=0, le=100, description="第三笔款项支付比例")
     payment_node_4: str | None = Field(None, max_length=100, description="第四笔款项支付节点")
-    payment_ratio_4: float | None = Field(None, ge=0, le=100, description="第四笔款项支付比例")
+    payment_ratio_4: Decimal | None = Field(None, ge=0, le=100, description="第四笔款项支付比例")
 
     # 软装费用
-    soft_budget: float | None = Field(None, description="软装预算金额")
-    soft_actual_cost: float | None = Field(None, description="软装实际发生成本")
+    soft_budget: Decimal | None = Field(None, description="软装预算金额")
+    soft_actual_cost: Decimal | None = Field(None, description="软装实际发生成本")
     soft_detail_attachment: str | None = Field(None, max_length=500, description="软装明细附件")
 
     # 其他费用
-    design_fee: float | None = Field(None, description="设计费用")
-    demolition_fee: float | None = Field(None, description="拆旧费用")
-    garbage_fee: float | None = Field(None, description="垃圾清运费用")
-    other_extra_fee: float | None = Field(None, description="其他额外费用")
+    design_fee: Decimal | None = Field(None, description="设计费用")
+    demolition_fee: Decimal | None = Field(None, description="拆旧费用")
+    garbage_fee: Decimal | None = Field(None, description="垃圾清运费用")
+    other_extra_fee: Decimal | None = Field(None, description="其他额外费用")
     other_fee_reason: str | None = Field(None, description="其他费用原因")
 
     model_config = ConfigDict(from_attributes=True)
@@ -110,22 +110,22 @@ class RenovationContractResponse(BaseModel):
     contract_end_date: datetime | None = None
     actual_start_date: datetime | None = None
     actual_end_date: datetime | None = None
-    hard_contract_amount: float | None = None
+    hard_contract_amount: Decimal | None = None
     payment_node_1: str | None = None
-    payment_ratio_1: float | None = None
+    payment_ratio_1: Decimal | None = None
     payment_node_2: str | None = None
-    payment_ratio_2: float | None = None
+    payment_ratio_2: Decimal | None = None
     payment_node_3: str | None = None
-    payment_ratio_3: float | None = None
+    payment_ratio_3: Decimal | None = None
     payment_node_4: str | None = None
-    payment_ratio_4: float | None = None
-    soft_budget: float | None = None
-    soft_actual_cost: float | None = None
+    payment_ratio_4: Decimal | None = None
+    soft_budget: Decimal | None = None
+    soft_actual_cost: Decimal | None = None
     soft_detail_attachment: str | None = None
-    design_fee: float | None = None
-    demolition_fee: float | None = None
-    garbage_fee: float | None = None
-    other_extra_fee: float | None = None
+    design_fee: Decimal | None = None
+    demolition_fee: Decimal | None = None
+    garbage_fee: Decimal | None = None
+    other_extra_fee: Decimal | None = None
     other_fee_reason: str | None = None
     created_at: datetime
     updated_at: datetime
